requirements.txt
Django==5.2.5
psycopg==3.2.9
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-environ==0.12.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.8.1
pytest==8.4.1
pytest-django==4.11.1